All files ziped
This includes every file required to open the project in visual studio, there are many zips to fit it under the file size restriction of github. they will be organised in order in folders on the repo. simply unzip them in folders of the same name on your own pc. the total file size is around 500mb so make sure you have enough room.